refactor affinity.go (#21)
* rename labels to LabelRule


* rename GroupInfo to runtimeGroupInfo


* refactor affinity.go


* unified RangeModification and keyRange


---------
