+++ pkg/schedule/affinity/labeler_test.go
@@ -1,0 +1,251 @@
+// Copyright 2025 TiKV Project Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package affinity
+
+import (
+	"context"
+	"testing"
+	"time"
+
+	"github.com/stretchr/testify/require"
+
+	"github.com/pingcap/kvproto/pkg/metapb"
+
+	"github.com/tikv/pd/pkg/core"
+	"github.com/tikv/pd/pkg/mock/mockconfig"
+	"github.com/tikv/pd/pkg/schedule/labeler"
+	"github.com/tikv/pd/pkg/storage"
+	"github.com/tikv/pd/pkg/utils/keyutil"
+)
+
+func TestKeyRangeOverlapValidation(t *testing.T) {
+	re := require.New(t)
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	store := storage.NewStorageWithMemoryBackend()
+	storeInfos := core.NewStoresInfo()
+	store1 := core.NewStoreInfo(&metapb.Store{Id: 1, Address: "test1"})
+	store1 = store1.Clone(core.SetLastHeartbeatTS(time.Now()))
+	storeInfos.PutStore(store1)
+
+	conf := mockconfig.NewTestOptions()
+
+	// Create manager without region labeler for basic validation testing
+	manager := NewManager(ctx, store, storeInfos, conf, nil)
+	err := manager.Initialize()
+	re.NoError(err)
+
+	validate := func(ranges []keyRange) error {
+		manager.Lock()
+		defer manager.Unlock()
+		return manager.validateNoKeyRangeOverlap(ranges)
+	}
+
+	// Test 1: Non-overlapping ranges should succeed
+	keyRanges1 := []keyRange{
+		{StartKey: []byte("a"), EndKey: []byte("b"), GroupID: "group1"},
+		{StartKey: []byte("c"), EndKey: []byte("d"), GroupID: "group1"},
+	}
+	err = validate(keyRanges1)
+	re.NoError(err, "Non-overlapping ranges should pass validation")
+
+	// Test 2: Overlapping ranges within same request should fail
+	keyRanges2 := []keyRange{
+		{StartKey: []byte("a"), EndKey: []byte("c"), GroupID: "group1"},
+		{StartKey: []byte("b"), EndKey: []byte("d"), GroupID: "group1"},
+	}
+	err = validate(keyRanges2)
+	re.Error(err, "Overlapping ranges should fail validation")
+	re.Contains(err.Error(), "overlap")
+
+	// Test 3: Adjacent ranges (not overlapping) should succeed
+	keyRanges3 := []keyRange{
+		{StartKey: []byte("a"), EndKey: []byte("b"), GroupID: "group1"},
+		{StartKey: []byte("b"), EndKey: []byte("c"), GroupID: "group1"},
+	}
+	err = validate(keyRanges3)
+	re.NoError(err, "Adjacent ranges should pass validation")
+
+	// Test 4: Verify checkKeyRangesOverlap function directly
+	overlaps := checkKeyRangesOverlap([]byte("a"), []byte("c"), []byte("b"), []byte("d"))
+	re.True(overlaps, "Ranges [a,c) and [b,d) should overlap")
+
+	overlaps = checkKeyRangesOverlap([]byte("a"), []byte("b"), []byte("c"), []byte("d"))
+	re.False(overlaps, "Ranges [a,b) and [c,d) should not overlap")
+}
+
+// TestKeyRangeOverlapRebuild tests rebuild after restart
+// Note: Full labeler integration test requires proper JSON serialization handling
+func TestKeyRangeOverlapRebuild(t *testing.T) {
+	re := require.New(t)
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	store := storage.NewStorageWithMemoryBackend()
+	storeInfos := core.NewStoresInfo()
+	store1 := core.NewStoreInfo(&metapb.Store{Id: 1, Address: "test1"})
+	store1 = store1.Clone(core.SetLastHeartbeatTS(time.Now()))
+	storeInfos.PutStore(store1)
+
+	conf := mockconfig.NewTestOptions()
+
+	manager := NewManager(ctx, store, storeInfos, conf, nil)
+	err := manager.Initialize()
+	re.NoError(err)
+
+	// Create two groups without key ranges for basic testing
+	groupsWithRanges := []GroupWithRanges{
+		{
+			Group: &Group{
+				ID:            "group1",
+				LeaderStoreID: 1,
+				VoterStoreIDs: []uint64{1},
+			},
+		},
+		{
+			Group: &Group{
+				ID:            "group2",
+				LeaderStoreID: 1,
+				VoterStoreIDs: []uint64{1},
+			},
+		},
+	}
+	err = manager.SaveAffinityGroups(groupsWithRanges)
+	re.NoError(err)
+
+	// Verify groups were created
+	re.True(manager.IsGroupExist("group1"))
+	re.True(manager.IsGroupExist("group2"))
+
+	// Create a new manager to simulate restart
+	manager2 := NewManager(ctx, store, storeInfos, conf, nil)
+	err = manager2.Initialize()
+	re.NoError(err)
+
+	// Verify groups were loaded from storage
+	re.True(manager2.IsGroupExist("group1"))
+	re.True(manager2.IsGroupExist("group2"))
+}
+
+func TestParseKeyRangesFromDataInvalidHex(t *testing.T) {
+	re := require.New(t)
+	_, err := parseKeyRangesFromData([]*labeler.KeyRangeRule{
+		{StartKeyHex: "zz", EndKeyHex: "10"},
+	}, "g1")
+	re.Error(err)
+	re.ErrorContains(err, "invalid hex start key")
+}
+
+func TestAffinityPersistenceWithLabeler(t *testing.T) {
+	re := require.New(t)
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	store := storage.NewStorageWithMemoryBackend()
+	storeInfos := core.NewStoresInfo()
+	store1 := core.NewStoreInfo(&metapb.Store{Id: 1, Address: "test1"})
+	store1 = store1.Clone(core.SetLastHeartbeatTS(time.Now()))
+	storeInfos.PutStore(store1)
+
+	conf := mockconfig.NewTestOptions()
+
+	regionLabeler, err := labeler.NewRegionLabeler(ctx, store, time.Second*5)
+	re.NoError(err)
+
+	manager := NewManager(ctx, store, storeInfos, conf, regionLabeler)
+	re.NoError(manager.Initialize())
+
+	gwr := GroupWithRanges{
+		Group: &Group{
+			ID:            "persist",
+			LeaderStoreID: 1,
+			VoterStoreIDs: []uint64{1},
+		},
+		KeyRanges: []keyutil.KeyRange{{StartKey: []byte{0x00}, EndKey: []byte{0x10}}},
+	}
+	re.NoError(manager.SaveAffinityGroups([]GroupWithRanges{gwr}))
+
+	// RangeCount should be recorded and label rule created.
+	state := manager.GetAffinityGroupState("persist")
+	re.NotNil(state)
+	re.Equal(1, state.RangeCount)
+	re.NotNil(regionLabeler.GetLabelRule(GetLabelRuleID("persist")))
+
+	// Reload manager to verify persistence and loadRegionLabel integration.
+	manager2 := NewManager(ctx, store, storeInfos, conf, regionLabeler)
+	re.NoError(manager2.Initialize())
+	state2 := manager2.GetAffinityGroupState("persist")
+	re.NotNil(state2)
+	re.Equal(1, state2.RangeCount)
+
+	// Remove all ranges and ensure cache/label are cleared.
+	ranges := []keyRange{{
+		StartKey: []byte{0x00},
+		EndKey:   []byte{0x10},
+		GroupID:  "persist",
+	}}
+	re.NoError(manager2.updateGroupRanges("persist", ranges))
+	re.NoError(manager2.updateGroupRanges("persist", nil))
+	state3 := manager2.GetAffinityGroupState("persist")
+	re.NotNil(state3)
+	re.Equal(0, state3.RangeCount)
+	re.Nil(regionLabeler.GetLabelRule(GetLabelRuleID("persist")))
+}
+
+// TestLabelRuleIntegration tests basic label rule creation and deletion
+// Note: Full integration with JSON serialization requires additional handling
+func TestLabelRuleIntegration(t *testing.T) {
+	re := require.New(t)
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	store := storage.NewStorageWithMemoryBackend()
+	storeInfos := core.NewStoresInfo()
+	store1 := core.NewStoreInfo(&metapb.Store{Id: 1, Address: "test1"})
+	store1 = store1.Clone(core.SetLastHeartbeatTS(time.Now()))
+	storeInfos.PutStore(store1)
+
+	conf := mockconfig.NewTestOptions()
+
+	// Create region labeler
+	regionLabeler, err := labeler.NewRegionLabeler(ctx, store, time.Second*5)
+	re.NoError(err)
+
+	manager := NewManager(ctx, store, storeInfos, conf, regionLabeler)
+	err = manager.Initialize()
+	re.NoError(err)
+
+	// Test: Group with no key ranges should not create label rule
+	group1 := &Group{
+		ID:            "group_no_label",
+		LeaderStoreID: 1,
+		VoterStoreIDs: []uint64{1},
+	}
+	err = manager.SaveAffinityGroups([]GroupWithRanges{{Group: group1}})
+	re.NoError(err)
+
+	labelRuleID := GetLabelRuleID("group_no_label")
+	labelRule := regionLabeler.GetLabelRule(labelRuleID)
+	re.Nil(labelRule, "No label rule should be created when keyRanges is nil")
+
+	// Verify group was created
+	re.True(manager.IsGroupExist("group_no_label"))
+
+	// Delete group (no key ranges, so force=false should work)
+	err = manager.DeleteAffinityGroup("group_no_label", false)
+	re.NoError(err)
+	re.False(manager.IsGroupExist("group_no_label"))
+}
